refactor: switch blocks content generators from PostMigrator to Post migration logic
